Fixes asshats and gets rid of grille electrocution (#15)
* Cluwnes, curse of cluwnes, and catbans

Adds cluwnes, cluwneing. Also adds curse of cluwnes and plasma fist
scroll to wizard. Also adds catbans. Cluwnes are a WIP - currently, when
gibbed or deleted their organs disappear (or shoot out when gibbed) and
the body falls to the ground with no organs. The body cannot be deleted
or player panel'd. The clowns have sounds and icons though, which is
nice, and the cluwning and cluwne curse spells work. Also, catbans
appear to work, which is nice. Plasma fist scroll works too which is
nice. Someone should take a look at the cluwnes and see what's wrong.

* got rid of the cluwneban shit

also made it setbrainloss instead of adjustbrainloss

* fixed comment suggestions

* Fixes asshats and gets rid of grille lightning bolts

Added a sprite for putting a butt on your head, and got rid of the 50%
chance to electrocute everyone in the area when an item is thrown at a
grille.